content/media/test/test_bug448534.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=448534
-->

<head>
  <title>Test for Bug 448534</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=448535">Mozilla Bug 448534</a>
<pre id="test">
<script class="testbody" type="text/javascript">

var videos = [];

function loaded(event) {
  var v = event.target;
  if (v._finished)
    return false;
  v.play();
  return false;
}

function started(event) {
  var v = event.target;
  if (v._finished)
    return false;
  ok(!v.paused, "Video should not be paused while playing");
  v.parentNode.removeChild(v);
  v._played = true;
  return false;
}

function stopped(event) {
  var v = event.target;
  if (v._finished)
    return false;
  v._finished = true;
  ok(v.paused, "Video should be paused after removing from the Document");
  if (AllFinished(videos))
    SimpleTest.finish();
  return false;
}


for (var i=0; i<gSmallTests.length; ++i) {
  var v = document.createElement('video');
  var test = gSmallTests[i];
  if (!v.canPlayType(test.type))
    continue;
  v.src = test.name;
  v._played = false;
  v._finished = false;
  v.addEventListener("loadedmetadata", loaded, false);
  v.addEventListener("play", started, false);
  v.addEventListener("pause", stopped, false);
  document.body.appendChild(v);
  videos.push(v);
}

if (videos.length == 0) {
  todo(false, "No types supported");
} else {
  SimpleTest.waitForExplicitFinish();
}

</script>
</pre>
</body>
</html>
